requirements.txt
kafka-python
psycopg2-binary
pyspark
lasio
pandas
plotly
dash
dash-core-components
dash-html-components
dash-table
